+++ src/test/java/blackjack/domain/betting/PlayerBettingTest.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.betting;
 
 import static org.assertj.core.api.Assertions.assertThat;
